refactor: migrate threeRenderer.service and remove afterwards unused isWhiteBackground.service
ref #2318
