+++ visualization/app/codeCharta/state/injector.service.ts
@@ -4,7 +4,6 @@ import { IsAttributeSideBarVisibleService } from "./store/appSettings/isAttribut
 import { FilesService } from "./store/files/files.service"
 import { IsLoadingFileService } from "./store/appSettings/isLoadingFile/isLoadingFile.service"
 import { MapSizeService } from "./store/treeMap/mapSize/mapSize.service"
-import { IsWhiteBackgroundService } from "./store/appSettings/isWhiteBackground/isWhiteBackground.service"
 import { ScalingService } from "./store/appSettings/scaling/scaling.service"
 import { MarkedPackagesService } from "./store/fileSettings/markedPackages/markedPackages.service"
 import { EdgesService } from "./store/fileSettings/edges/edges.service"
@@ -22,7 +21,6 @@ export class InjectorService {
 		private isLoadingFileService: IsLoadingFileService,
 		private filesService: FilesService,
 		private mapSizeService: MapSizeService,
-		private isWhiteBackgroundService: IsWhiteBackgroundService,
 		private scalingService: ScalingService,
 		private markedPackagesService: MarkedPackagesService,
 		private edgesService: EdgesService,
