+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeViewerService.ts
@@ -1,7 +1,7 @@
 import { ThreeSceneService } from "./threeSceneService"
 import { ThreeCameraService } from "./threeCameraService"
 import { ThreeOrbitControlsService } from "./threeOrbitControlsService"
-import { ThreeRendererService } from "./threeRendererService"
+import { ThreeRendererService } from "./threeRenderer.service"
 import { ThreeStatsService } from "./threeStatsService"
 
 export class ThreeViewerService {
