+++ visualization/mocks/ng.mockhelper.ts
@@ -13,6 +13,9 @@ export function instantiateModule(id: string) {
 	angular.mock.module(($provide: ng.auto.IProvideService) => {
 		$provide.value("viewCubeMouseEvents", {})
 	})
+	angular.mock.module(($provide: ng.auto.IProvideService) => {
+		$provide.value("threeRendererService", { render: jest.fn() })
+	})
 }
 
 export function getService<T>(id: string): T {
